+++ out/src/libs/stringFormatter.js
@@ -3,17 +3,19 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.whiteSpaceRemove = void 0;
 function whiteSpaceRemove(str, spaceSize) {
     str = str.trim();
-    var identationSpace = str === ''
-        ? ''
-        : ' '.repeat(spaceSize);
-    return identationSpace + str.replace(/([^"]+)|("[^"]+")/g, function ($0, $1, $2) {
-        if ($1) {
-            return $1.replace(/\s{1,}/g, " ");
-        }
-        else {
-            return $2;
-        }
-    });
+    var identationSpace = str === "" ? "" : " ".repeat(spaceSize);
+    return (identationSpace +
+        str
+            .replace(/,/g, ", ")
+            .replace(/([^"]+)|("[^"]+")/g, function ($0, $1, $2) {
+            if ($1) {
+                return $1.replace(/\s{1,}/g, " ");
+            }
+            else {
+                return $2;
+            }
+        })
+            .trim());
 }
 exports.whiteSpaceRemove = whiteSpaceRemove;
 //# sourceMappingURL=stringFormatter.js.map