+++ ChangeLog.md
@@ -4,6 +4,37 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.2] - 2021-11-27
+
+Create some regex to formatte document.
+
+### Added
+
+### Changed
+
+### Fixed
+- [Add formatter to rest of types](https://github.com/AndreOneti/qml-formatter/issues/1)
+
+Create more regex to formatte some files.
+
+Can't format if has "if/else" esteatement.
+
+------
+
+## [1.0.1] - 2021-11-26
+
+Fix import format.
+
+### Added
+
+### Changed
+
+### Fixed
+- [qml-formatter](https://github.com/AndreOneti/qml-formatter)
+  Fix "import" formatter
+
+------
+
 ## [1.0.0] - 2021-11-26
 
 Create a firt formater patter, init with import.